BACKEND/config.py
import os
from dotenv import load_dotenv

load_dotenv()  # Load .env variables

GROQ_API_KEY = os.getenv("GROQ_API_KEY")
GROQ_API_BASE = "https://api.groq.com/openai/v1"
GROQ_MODEL = "llama3-8b-8192" # text to text  , code  , no audio  , video or image , no web searches
SERPAPI_KEY = os.getenv("SERPAPI_KEY")
EMBEDDING_MODEL = "sentence-transformers/all-MiniLM-L6-v2"
DB_PATH = "vectorstore/db_faiss"

SYSTEM_PROMPT = """
You are an intent classification assistant.
Classify the user query into one or more of these categories:

- weather        → Any query about weather, temperature, forecast, climate
- local_rag      → Any query that could be answered from internal/local documents
- web_search     → Queries that need fresh, external, or general web information

Rules:
- If the query is small talk, greetings, or unrelated to these categories, return an empty list []
- Only choose from: ["weather", "local_rag", "web_search"]
- Output must be a valid JSON array with no extra text.

Examples:
User query: "hi"
Intents: []

User query: "hello there"
Intents: []

User query: "how are you?"
Intents: []

User query: "what's the weather in Bengaluru?"
Intents: ["weather"]

User query: "summarize our sales report"
Intents: ["local_rag"]

User query: "latest news on smart doorbells"
Intents: ["web_search"]

User query: "do we have documents on doorbells and also latest news?"
Intents: ["local_rag", "web_search"]
"""

