BACKEND/requirements.txt
Flask==3.1.1
flask_cors==6.0.1
google_search_results==2.4.2
langchain==0.3.27
langchain_community==0.3.27
openai==1.99.9
python-dotenv==1.1.1
Requests==2.32.4
trafilatura==2.0.0
Werkzeug==3.1.3
